M1/azor-chatdog-py/requirements.txt
google-genai
python-dotenv
llama-cpp-python
ollama
anthropic
prompt_toolkit
colorama
pydantic>=2.0.0
fpdf2
markdown
